requirements.txt
crewai[litellm]
python-dotenv
requests
icalendar
python-dateutil
twilio
google-genai
